refactor(status): extract EpicDependencyService to reduce LOC
Extract epic dependency status logic into dedicated service module
to comply with LOC limits (status_query_service.py: 608 → 475 lines).

Changes:
- Create src/vibe3/services/epic_dependency_service.py (160 lines)
- Move _parse_dependencies_from_body() function
- Move check_epic_dependency_status() and enrich_epic_items() methods
- Add delegation methods in StatusQueryService
- Update test imports to use new module

Verification:
- All 28 status_query_utils tests pass
- All 16 task_status tests pass
- mypy type check passes
- File LOC now under 480 limit (475 lines)

Ref: #1621
