+++ tests/vibe3/services/test_status_query_utils.py
@@ -3,9 +3,9 @@
 from unittest.mock import MagicMock
 
 from vibe3.models.orchestration import IssueState
+from vibe3.services.epic_dependency_service import _parse_dependencies_from_body
 from vibe3.services.status_query_service import (
     StatusQueryService,
-    _parse_dependencies_from_body,
     is_auto_task_branch,
     is_canonical_task_branch,
     issue_priority,
